shared: Fix resource class
